Refine diagnostics panel visibility
